Conflicts:
	config/bascs.conf
	config/victoria.conf
	export/maestro.csv
	imagenes/AATECH18-CI-48CHR1.jpg
	nbproject/genfiles.properties
	nbproject/project.properties
	nbproject/project.xml
	src/etl/bascs/impala/clases/Producto.java
	src/etl/bascs/impala/clases/ProductosVictoria.java
	src/etl/bascs/impala/config/Propiedades.java
	src/etl/bascs/impala/json/ConsultaHttp.java
	src/etl/bascs/impala/json/ConsultaHttpVictoria.java
	src/etl/bascs/impala/main.form
	src/etl/bascs/impala/main.java
	src/etl/bascs/impala/worker/DetalleWorker.java
	src/etl/bascs/impala/worker/MaestroWorker.java
	src/etl/bascs/impala/worker/PrestashopWorker.java

+++ src/bascs/website/clases/CuotasSC.java
@@ -14,8 +14,8 @@ public class CuotasSC {
     private Integer producto_id;
     private Integer cuotas;
     private Integer posee_descuento;
-    private Double importe_cuota;
-    private Double porcentaje_descuento;
+    private Integer importe_cuota;
+    private Integer porcentaje_descuento;
 
     public Integer getId() {
         return id;
@@ -49,19 +49,19 @@ public class CuotasSC {
         this.posee_descuento = posee_descuento;
     }
 
-    public Double getImporte_cuota() {
+    public Integer getImporte_cuota() {
         return importe_cuota;
     }
 
-    public void setImporte_cuota(Double importe_cuota) {
+    public void setImporte_cuota(Integer importe_cuota) {
         this.importe_cuota = importe_cuota;
     }
 
-    public Double getPorcentaje_descuento() {
+    public Integer getPorcentaje_descuento() {
         return porcentaje_descuento;
     }
 
-    public void setPorcentaje_descuento(Double porcentaje_descuento) {
+    public void setPorcentaje_descuento(Integer porcentaje_descuento) {
         this.porcentaje_descuento = porcentaje_descuento;
     }
